chore: move headfoot out of util
